Account-klassen er fjernet. Funktionaliteten er blevet flyttet til Player.
