+++ src/main/Spil/Model/Player.java
@@ -1,7 +1,5 @@
 package main.Spil.Model;
 
-import main.Spil.Model.Account;
-
 /**
  * This class is for creating the players.
  */
@@ -10,18 +8,20 @@ public class Player {
      * Defines the name of the player.
      */
     private String name;
-    /**
-     * Defines the Account as account.
-     */
-    public Account account;
 
     /**
-     * This method is setting a name for the player and creating an account for the player.
+     * Defines the balance.
+     */
+    private int balance = 0;
+
+
+    /**
+     * This method is setting a name for the player and creating balance for the player.
      * @param name Is used for defining the name of the player.
      */
     public Player(String name){
         this.name = name;
-        account = new Account(1000);
+        this.balance = balance;
     }
 
     /**
@@ -31,4 +31,24 @@ public class Player {
     public String getName() {
         return name;
     }
+
+
+    /**
+     * This method is changing the balance with a specific value.
+     *
+     * @param amount Is a specific value to change the balance with.
+     */
+    public void changeBalance(int amount) {
+        if (balance < 0) this.balance = 0;
+        balance += amount;
+    }
+
+    /**
+     * @return Returns the balance
+     */
+    public int getBalance() {
+        return balance;
+    }
+
+
 }
